requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-ninja==1.2.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
